Correct indexing behavior.
When creating a new object store, the user will expect that the index that they
define will correspond directly with the keypath of the indexable values that
they pass as entries in the object store.  Previously, the user had to include
`$data...` at the beginning of the indexes - both in the object store
definition and when querying.  This was counter-intuitive.

At a later time, it may be worth supporting `$meta...` indexes - for example,
so that a user could filter results by which entries had ever been modified.